+++ src/main/java/xyz/kohara/stellarity/interface_injection/ExtEndSpike.java
@@ -1,6 +1,8 @@
 package xyz.kohara.stellarity.interface_injection;
 
 
+import net.minecraft.world.level.levelgen.feature.SpikeFeature;
+
 public interface ExtEndSpike {
     default boolean stellarity$hasAltar() {
         throw new AssertionError("Not transformed");
@@ -8,5 +10,15 @@ public interface ExtEndSpike {
 
     default void stellarity$setAltar(boolean hasAltar) {
         throw new AssertionError("Not transformed");
+    }
+
+    static SpikeFeature.EndSpike apply(SpikeFeature.EndSpike spike, boolean hasAltar) {
+        spike.stellarity$setAltar(hasAltar);
+        return spike;
+    }
+
+    static SpikeFeature.EndSpike applyDefaults(SpikeFeature.EndSpike spike) {
+        spike.stellarity$setAltar(false);
+        return spike;
     }
 }